+++ postprocessor-service/src/main/java/io/pockethive/postprocessor/Application.java
@@ -1,5 +1,6 @@
 package io.pockethive.postprocessor;
 
+import io.pockethive.sink.clickhouse.ClickHouseSinkProperties;
 import io.pockethive.worker.sdk.metrics.PrometheusPushGatewayProperties;
 import org.springframework.amqp.rabbit.annotation.EnableRabbit;
 import org.springframework.boot.SpringApplication;
@@ -10,7 +11,7 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @EnableRabbit
 @EnableScheduling
 @SpringBootApplication
-@EnableConfigurationProperties(PrometheusPushGatewayProperties.class)
+@EnableConfigurationProperties({PrometheusPushGatewayProperties.class, ClickHouseSinkProperties.class})
 public class Application {
 
   public static void main(String[] args) {